+++ app/domain/token_factory.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'date'
+
 class TokenFactory < Dry::Struct
 
   NoSigningKey = ::Util::ErrorClass.new(
@@ -7,12 +9,28 @@ class TokenFactory < Dry::Struct
 
   attribute :slosilo, ::Types::Any.default{ Slosilo }
 
+  MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION = 5.hours
+
   def signing_key(account)
     slosilo["authn:#{account}".to_sym] || raise(NoSigningKey, account)
   end
-    
-  def signed_token(account:, username:)
-    signing_key(account).issue_jwt(sub: username)
+
+  def signed_token(account:,
+                   username:,
+                   host_ttl: Rails.application.config.conjur_config.host_authorization_token_ttl,
+                   user_ttl: Rails.application.config.conjur_config.user_authorization_token_ttl)
+    signing_key(account).issue_jwt(
+      sub: username,
+      exp: Time.now + offset(
+        ttl: username.starts_with?('host/') ? host_ttl : user_ttl
+      )
+    )
+  end
+
+  def offset(ttl:)
+    return ttl.to_i if ttl.to_i < MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION
+
+    MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION
   end
 
 end